split full_test and remove full_test
